Update associations in classes
