+++ book.rb
@@ -1,0 +1,17 @@
+# Represents a book with attributes such as title, author, and rentals.
+class Book
+  attr_accessor :title, :author
+  attr_reader :rentals
+
+  def initialize(title, author)
+    @title = title
+    @author = author
+    @rentals = []
+  end
+
+  def add_rental(person, date)
+    rental = Rental.new(date, self, person)
+    @rentals << rental
+    person.rentals << rental
+  end
+end
