+++ student.rb
@@ -1,8 +1,11 @@
 require_relative 'person'
+
 # Represents a sub-class of Person.
 class Student < Person
+  attr_accessor :classroom
+
   def initialize(age, classroom:, parent_permission: true, name: 'Unknown')
-    super(age, parent_permission: parent_permission, name: name)
+    super(age, name: name, parent_permission: parent_permission)
     @classroom = classroom
   end
